+++ app/javascript/components/v1/HeaderPage.vue
@@ -59,7 +59,7 @@ watch(Hambarg as Ref<boolean>, (newValue: boolean) => {
             </li>
           </div>
           <li class="cursor-pointer text-white hover:text-slate-300 duration-500 pb-2 font-black text-lg animate-gnaviAnime animate-duration-1000 animate-delay-200 animate-fill-forwards opacity-0">
-            <RouterLink to="/Login" class="block w-24 text-base md:text-xl">ログイン</RouterLink>
+            <RouterLink to="/Login" class="block w-24 text-base md:text-xl" @click="Hambarg = !Hambarg">ログイン</RouterLink>
           </li>
         </ul>
       </div>
