+++ app/models/user.rb
@@ -2,14 +2,14 @@ class User < ApplicationRecord
   before_save {self.email = email.downcase}
 
   #名前のバリデーション
-  validates :name, presence :true, length: {maximum:25}
+  validates :name, presence: true, length: {maximum:25}
 
   #メールのバリデーション
   EMAIL_VALIDATE = /\A[\w+\-.]+@[a-z\d\-.]+\.[a-z]+\z/i
   validates :email, presence: true,
                     uniqueness: { case_sensitive: false },
                     length: { maximum: 105 },
-                    format: { with: VALID_EMAIL_REGEX }
+                    format: { with: EMAIL_VALIDATE }
 
   #パスワードのバリデーション
   has_secure_password
